updated: move up UseCase public header files
