+++ src/Usecase/CMakeLists.txt
@@ -4,12 +4,12 @@ add_library(vgg_usecase STATIC
 )
 
 target_include_directories(vgg_usecase PUBLIC
-  ${CMAKE_CURRENT_SOURCE_DIR}/include
+  ${CMAKE_SOURCE_DIR}/include
 )
 
 target_include_directories(vgg_usecase PRIVATE
-  ${CMAKE_CURRENT_SOURCE_DIR}/include/Usecase
   ${CMAKE_CURRENT_SOURCE_DIR}
+  ${CMAKE_SOURCE_DIR}/include/Usecase
 )
 
 target_link_libraries(vgg_usecase PUBLIC
